Move logic from TypeExtensions to DefaultNamingConventions
